requirements.txt
anyio==4.9.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.4.26
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
python-telegram-bot==13.15
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tornado==6.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
